added the facebook helper to clean up loginView
